+++ platform_device_id/example/macos/.gitignore
@@ -3,4 +3,5 @@
 **/Pods/
 
 # Xcode-related
+**/dgph
 **/xcuserdata/
