+++ platform_device_id/example/macos/Runner/Base.lproj/MainMenu.xib
@@ -323,6 +323,10 @@
                         </items>
                     </menu>
                 </menuItem>
+                <menuItem title="Help" id="EPT-qC-fAb">
+                    <modifierMask key="keyEquivalentModifierMask"/>
+                    <menu key="submenu" title="Help" systemMenu="help" id="rJ0-wn-3NY"/>
+                </menuItem>
             </items>
             <point key="canvasLocation" x="142" y="-258"/>
         </menu>
