+++ platform_device_id/example/macos/Runner/Configs/AppInfo.xcconfig
@@ -5,10 +5,10 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = example
+PRODUCT_NAME = platform_device_id_example
 
 // The application's bundle identifier
-PRODUCT_BUNDLE_IDENTIFIER = com.di1shuai.example
+PRODUCT_BUNDLE_IDENTIFIER = com.example.platformDeviceIdExample
 
 // The copyright displayed in application information
-PRODUCT_COPYRIGHT = Copyright © 2021 com.di1shuai. All rights reserved.
+PRODUCT_COPYRIGHT = Copyright © 2025 com.example. All rights reserved.
